refactor(app): split sql modal reducers
