+++ src/app/update/mod.rs
@@ -23,4 +23,4 @@ pub use explain::reduce_explain;
 pub use explain::reduce_explain_with_services;
 pub use helpers::{char_to_byte_index, validate_all, validate_field};
 pub use modal::dispatch_modal;
-pub use sql_editor::reduce_sql_modal;
+pub use sql_editor::dispatch_sql_modal;
